requirements.txt
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-s3
aws-cdk.aws-s3-deployment
aws-cdk.aws-mwaa
aws-cdk.core

